Started replacepluswithspace program.
Intent of program is to call it anywhere from a terminal program (so I'll this spot to my PATH variable) and pass into it a relative or absolute filepath to rename a file if it has any '+' characters with the ' ' (space) character. I want to do this in C so that I can invoke this from anywhere instead of Python where I have to run the program by calling Python and specifying the full-path of the Python script. There's probably other workarounds but I thought having a C way to do this would be fun.